+++ infra/tests/test_monitoring_stack.py
@@ -1,7 +1,6 @@
-import json
 import os
 import unittest
-from typing import Any, cast
+from typing import Any
 from unittest import mock
 
 import aws_cdk as cdk
@@ -12,7 +11,6 @@ from aws_cdk import (
     aws_elasticache,
     aws_elasticloadbalancingv2 as aws_elb,
     aws_rds,
-    aws_s3,
 )
 
 from constants import (
@@ -24,10 +22,8 @@ from constants import (
     get_slack_notification_config,
 )
 from monitoring_stack import MonitoringStack
-from runtime_iam import create_tracker_task_role, create_worker_task_role
 from shared import SharedStack
 from stage import DEV, DEV_STACK_PREFIX, PROD, Stage
-from stage_config import ManagedAwsRuntimeConfig
 from tracker_stack import TrackerStack
 from worker_stack import WorkerStack
 
@@ -69,39 +65,6 @@ def _has_resource_property(
 
 def _has_logical_id_prefix(template: assertions.Template, resource_type: str, prefix: str) -> bool:
     return any(logical_id.startswith(prefix) for logical_id in template.find_resources(resource_type))
-
-
-def _named_role(template: assertions.Template, role_name: str) -> tuple[str, JsonObject]:
-    resources = cast(dict[str, JsonObject], template.find_resources("AWS::IAM::Role"))
-    matches = [
-        (logical_id, resource)
-        for logical_id, resource in resources.items()
-        if cast(JsonObject, resource.get("Properties", {})).get("RoleName") == role_name
-    ]
-    if len(matches) != 1:
-        raise AssertionError(f"expected one role named {role_name!r}, found {len(matches)}")
-    return matches[0]
-
-
-def _role_policy_statements(template: assertions.Template, role_logical_id: str) -> list[JsonObject]:
-    statements: list[JsonObject] = []
-    policies = cast(dict[str, JsonObject], template.find_resources("AWS::IAM::Policy"))
-    for policy in policies.values():
-        properties = cast(JsonObject, policy.get("Properties", {}))
-        if {"Ref": role_logical_id} not in cast(list[JsonObject], properties.get("Roles", [])):
-            continue
-        policy_document = cast(JsonObject, properties["PolicyDocument"])
-        policy_statements = policy_document["Statement"]
-        if isinstance(policy_statements, list):
-            statements.extend(cast(list[JsonObject], policy_statements))
-        else:
-            statements.append(cast(JsonObject, policy_statements))
-    return statements
-
-
-def _statement_actions(statement: JsonObject) -> set[str]:
-    actions = cast(str | list[str], statement["Action"])
-    return set(actions) if isinstance(actions, list) else {actions}
 
 
 def _monitoring_template(stage_name: str = PROD) -> assertions.Template:
@@ -185,7 +148,7 @@ def _shared_template(stage_name: str = PROD) -> assertions.Template:
     return assertions.Template.from_stack(shared)
 
 
-def _service_templates(stage_name: str) -> tuple[assertions.Template, assertions.Template]:
+def service_templates(stage_name: str) -> tuple[assertions.Template, assertions.Template]:
     app = cdk.App(context=SHARED_STACK_CONTEXT)
     stage = Stage(stage_name)
     env = cdk.Environment(account=TEST_AWS_ACCOUNT, region=TEST_AWS_REGION)
@@ -316,7 +279,7 @@ class MonitoringStackTest(unittest.TestCase):
 
     def test_dev_stage_wires_stage_config_to_resources(self) -> None:
         with mock.patch.dict(os.environ, TEST_DEV_ENV, clear=True):
-            tracker_template, worker_template = _service_templates(DEV)
+            tracker_template, worker_template = service_templates(DEV)
         with mock.patch.dict(os.environ, TEST_ALERTS_SLACK_ENV, clear=True):
             monitoring_template = _monitoring_template(DEV)
 
@@ -356,7 +319,7 @@ class MonitoringStackTest(unittest.TestCase):
         ):
             environment = TEST_DEV_ENV if stage_name == DEV else {}
             with self.subTest(stage=stage_name), mock.patch.dict(os.environ, environment, clear=True):
-                tracker_template, worker_template = _service_templates(stage_name)
+                tracker_template, worker_template = service_templates(stage_name)
 
                 expected_env = assertions.Match.array_with(
                     [
@@ -384,7 +347,7 @@ class MonitoringStackTest(unittest.TestCase):
 
     def test_dev_sentry_secret_is_optional(self) -> None:
         with mock.patch.dict(os.environ, TEST_DEV_ENV, clear=True):
-            tracker_template, worker_template = _service_templates(DEV)
+            tracker_template, worker_template = service_templates(DEV)
 
         self.assertNotIn("SENTRY_DSN", str(tracker_template.to_json()))
         self.assertNotIn("SENTRY_DSN", str(worker_template.to_json()))
@@ -395,7 +358,7 @@ class MonitoringStackTest(unittest.TestCase):
             "SENTRY_DSN_SECRET_NAME": custom_sentry_secret_name,
         }
         with mock.patch.dict(os.environ, sentry_environment, clear=True):
-            tracker_template, worker_template = _service_templates(DEV)
+            tracker_template, worker_template = service_templates(DEV)
 
         for template in (tracker_template, worker_template):
             template.has_resource_properties(
@@ -423,173 +386,6 @@ class MonitoringStackTest(unittest.TestCase):
             ]
             self.assertEqual(len(sentry_value_from), 1)
             self.assertIn(custom_sentry_secret_name, str(sentry_value_from[0]))
-
-    def test_managed_runtime_task_roles_are_scoped_and_closed_by_default(self) -> None:
-        with mock.patch.dict(os.environ, {}, clear=True):
-            tracker_template, worker_template = _service_templates(DEV)
-
-        expected_environment = assertions.Match.array_with(
-            [
-                {"Name": "AWS_DEPLOYMENT_ROLE_ORG_IDS", "Value": ""},
-                {"Name": "AWS_DEPLOYMENT_REGION", "Value": TEST_AWS_REGION},
-                assertions.Match.object_like({"Name": "AWS_DEPLOYMENT_S3_BUCKET"}),
-                {"Name": "AWS_DEPLOYMENT_LOG_GROUP", "Value": "/valkyrie/benchmarks-dev"},
-                {"Name": "AWS_DEPLOYMENT_LOG_RETENTION_DAYS", "Value": "7"},
-                {"Name": "AWS_MANAGED_SUBMISSIONS_ENABLED", "Value": "false"},
-            ]
-        )
-
-        expected_actions = {
-            "s3:ListBucket",
-            "s3:GetObject",
-            "s3:PutObject",
-        }
-        for template, role_name, output_name, service_actions in (
-            (
-                tracker_template,
-                "ValkyrieTrackerTaskRole-dev",
-                "TrackerTaskRoleArn",
-                expected_actions,
-            ),
-            (
-                worker_template,
-                "ValkyrieWorkerTaskRole-dev",
-                "WorkerTaskRoleArn",
-                expected_actions
-                | {
-                    "logs:CreateLogGroup",
-                    "logs:PutRetentionPolicy",
-                    "logs:CreateLogStream",
-                    "logs:PutLogEvents",
-                    "ecs:UpdateTaskProtection",
-                },
-            ),
-        ):
-            with self.subTest(role=role_name):
-                role_logical_id, _ = _named_role(template, role_name)
-                task_definitions = cast(
-                    dict[str, JsonObject],
-                    template.find_resources("AWS::ECS::TaskDefinition"),
-                )
-                self.assertEqual(len(task_definitions), 1)
-                task_properties = cast(JsonObject, next(iter(task_definitions.values()))["Properties"])
-                self.assertEqual(task_properties["TaskRoleArn"], {"Fn::GetAtt": [role_logical_id, "Arn"]})
-                self.assertIn("ExecutionRoleArn", task_properties)
-                self.assertNotEqual(task_properties["TaskRoleArn"], task_properties["ExecutionRoleArn"])
-                self.assertEqual(
-                    template.to_json()["Outputs"][output_name]["Value"],
-                    {"Fn::GetAtt": [role_logical_id, "Arn"]},
-                )
-                template.has_resource_properties(
-                    "AWS::ECS::TaskDefinition",
-                    {
-                        "ContainerDefinitions": assertions.Match.array_with(
-                            [assertions.Match.object_like({"Environment": expected_environment})]
-                        )
-                    },
-                )
-
-                statements = _role_policy_statements(template, role_logical_id)
-                actions = set[str]().union(*(_statement_actions(statement) for statement in statements))
-                self.assertEqual(actions, service_actions)
-                self.assertFalse(
-                    actions
-                    & {
-                        "s3:DeleteObject",
-                        "sts:AssumeRole",
-                        "secretsmanager:GetSecretValue",
-                        "lambda:InvokeFunction",
-                        "kms:Decrypt",
-                        "kms:GenerateDataKey",
-                    }
-                )
-
-                list_statement = next(
-                    statement for statement in statements if _statement_actions(statement) == {"s3:ListBucket"}
-                )
-                self.assertEqual(
-                    list_statement["Condition"],
-                    {"StringLike": {"s3:prefix": ["agents/*", "benchmarks/*"]}},
-                )
-                get_statement = next(
-                    statement for statement in statements if _statement_actions(statement) == {"s3:GetObject"}
-                )
-                self.assertIn("agents/*", json.dumps(get_statement["Resource"]))
-                self.assertIn("benchmarks/*", json.dumps(get_statement["Resource"]))
-                put_statement = next(
-                    statement for statement in statements if _statement_actions(statement) == {"s3:PutObject"}
-                )
-                self.assertIn("benchmarks/*", json.dumps(put_statement["Resource"]))
-                self.assertNotIn("agents/*", json.dumps(put_statement["Resource"]))
-
-                for statement in statements:
-                    resources = statement["Resource"]
-                    if resources == "*" or (isinstance(resources, list) and "*" in resources):
-                        self.assertEqual(_statement_actions(statement), {"ecs:UpdateTaskProtection"})
-
-                if role_name.startswith("ValkyrieWorker"):
-                    log_statement = next(
-                        statement for statement in statements if "logs:CreateLogGroup" in _statement_actions(statement)
-                    )
-                    self.assertIn("/valkyrie/benchmarks-dev/*", json.dumps(log_statement["Resource"]))
-                    log_stream_statement = next(
-                        statement for statement in statements if "logs:PutLogEvents" in _statement_actions(statement)
-                    )
-                    self.assertIn(":log-stream:*", json.dumps(log_stream_statement["Resource"]))
-
-    def test_managed_runtime_optional_grants_are_limited_to_configured_resources(self) -> None:
-        app = cdk.App()
-        stack = cdk.Stack(
-            app,
-            "RuntimeIamStack",
-            env=cdk.Environment(account=TEST_AWS_ACCOUNT, region=TEST_AWS_REGION),
-        )
-        bucket = aws_s3.Bucket.from_bucket_name(stack, "ManagedRuntimeBucket", "managed-runtime-bucket")
-        kms_key_arn = f"arn:aws:kms:{TEST_AWS_REGION}:{TEST_AWS_ACCOUNT}:key/test-key"
-        config = ManagedAwsRuntimeConfig(
-            benchmark_log_group_prefix="/valkyrie/benchmarks",
-            benchmark_log_retention_days=7,
-            tracker_secret_name_prefixes=("valkyrie/tracker/",),
-            worker_secret_name_prefixes=("valkyrie/worker/",),
-            tracker_lambda_function_name_patterns=("valkyrie-analyzer-*",),
-            worker_lambda_function_name_patterns=("valkyrie-post-run-*",),
-            kms_key_arns=(kms_key_arn,),
-        )
-        create_tracker_task_role(stack, Stage(DEV), bucket, config)
-        create_worker_task_role(stack, Stage(DEV), bucket, config)
-        template = assertions.Template.from_stack(stack)
-
-        for role_name, secret_prefix, lambda_pattern in (
-            ("ValkyrieTrackerTaskRole-dev", "valkyrie/tracker/", "valkyrie-analyzer-*"),
-            ("ValkyrieWorkerTaskRole-dev", "valkyrie/worker/", "valkyrie-post-run-*"),
-        ):
-            with self.subTest(role=role_name):
-                role_logical_id, _ = _named_role(template, role_name)
-                statements = _role_policy_statements(template, role_logical_id)
-
-                secret_statement = next(
-                    statement
-                    for statement in statements
-                    if _statement_actions(statement) == {"secretsmanager:GetSecretValue"}
-                )
-                self.assertIn(f"secret:{secret_prefix}*", json.dumps(secret_statement["Resource"]))
-
-                lambda_statement = next(
-                    statement for statement in statements if _statement_actions(statement) == {"lambda:InvokeFunction"}
-                )
-                self.assertIn(f"function:{lambda_pattern}", json.dumps(lambda_statement["Resource"]))
-
-                kms_statement = next(
-                    statement for statement in statements if "kms:Decrypt" in _statement_actions(statement)
-                )
-                self.assertEqual(
-                    _statement_actions(kms_statement),
-                    {"kms:Decrypt", "kms:GenerateDataKey"},
-                )
-                self.assertEqual(kms_statement["Resource"], kms_key_arn)
-
-                for statement in (secret_statement, lambda_statement, kms_statement):
-                    self.assertNotEqual(statement["Resource"], "*")
 
 
 if __name__ == "__main__":
